Ginthi_Backend-development/ginthi_agents/auth_service/requirements.txt
# Requirements for the auth_service (grouped and annotated)


## Core runtime / Web framework
fastapi==0.117.1
fastapi-mcp==0.4.0  # FastAPI MCP integration
uvicorn==0.36.0  # ASGI server used to run the FastAPI app

## Authentication / Security (add as needed)


## Relational database / async
asyncpg==0.30.0  # Async PostgreSQL driver
psycopg2-binary==2.9.10  # PostgreSQL driver (sync) - sometimes required
sqlalchemy==2.0.43  # ORM / SQL toolkit

## Configuration & validation
python-dotenv==1.1.1  # load environment variables from .env
pydantic==2.11.9  # data validation and settings management
# Optional pydantic extras (email validators)
pydantic[email]

## HTTP clients & logging
requests==2.32.5  # HTTP client for external API calls
python-json-logger==4.0.0  # JSON formatter for structured logging

## Data processing & AWS
pandas  # Data manipulation and analysis
pyarrow  # Apache Arrow for efficient data serialization
boto3  # AWS SDK for Python

# Notes:
# - When deploying, use a pinned `requirements-prod.txt` for reproducible installs.
# - Create `requirements-dev.txt` for test and lint dependencies (pytest, black, flake8, isort).
# - If/when you add auth-specific packages, move them under the Authentication section.
# Authentication & Security
python-jose[cryptography]   # JWT handling (JOSE implementation)
passlib[bcrypt]             # Password hashing
python-multipart
